move init function to its own file
